Refactor Calendar Legend Component and Extract LegendIcon
- Moved the LegendIcon component from CalendarLegend.tsx to a new file LegendIcon.tsx for better modularity and readability.
- Updated CalendarLegend to import and utilize the new LegendIcon component, streamlining the code structure.
- The LegendIcon component handles rendering of color indicators for calendar items, supporting both mixed and single color states.

This refactor enhances maintainability and clarity in the Calendar UI components.
